+++ src/main/java/se/olapetersson/ReactiveApplication.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class ReactiveApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(ReactiveApplication.class, args);
     }
