+++ src/main/java/se/olapetersson/websocket/LobbyWebSocketController.java
@@ -1,6 +1,5 @@
 package se.olapetersson.websocket;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
@@ -12,14 +11,12 @@ import java.util.List;
 
 @Controller
 public class LobbyWebSocketController {
-
     @Autowired
     LobbyRepository repository;
 
     @MessageMapping("/lobbies/init")
     @SendTo("/topic/lobbies")
-    public List<Lobby> getLobbies(String ping){
-
+    public List<Lobby> getLobbies() {
         return repository.findAll();
     }
 }
