+++ src/main/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+    jpa:
+        generate-ddl: true
+        show-sql: true
+        hibernate:
+            ddl-auto: create-drop
+    datasource:
+        url: jdbc:h2:file:./blog
+    jackson:
+        date-format: yyyy-MM-dd'T'HH:mm:ss.SSSXXX
