+++ src/main/resources/static/css/my.css
@@ -1,5 +1,5 @@
 body {
-    background-color: cadetblue; !important;
+    background-color: cadetblue !important;
 }
 
 #container > .jumbotron {
