+++ src/main/webapp/index.html
@@ -2,46 +2,39 @@
 <html>
 <head>
     <title>Hello WebSocket</title>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
-    <script src="js/stomp.js"></script>
+    <script src="/webjars/jquery/2.0.0/jquery.min.js"></script>
+    <script src="/js/stomp.js"></script>
     <!-- bootstrap -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css" integrity="sha512-dTfge/zgoMYpP7QbHy4gWMEGsbsdZeCXz7irItjcC3sPUFtf0kuFbDz/ixG7ArTxmDjLXDmezHubeNikyKGVyQ==" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap-theme.min.css" integrity="sha384-aUGj/X2zp5rLCbBxumKTCw2Z50WgIr1vs/PFN4praOTvYXWlVyh2UtNUU0KAUhAX" crossorigin="anonymous">
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js" integrity="sha512-K1qjQ+NcF2TYO/eI3M6v8EiNYZfA95pQumfvcVrTHtwQVDG+aHRqLi/ETn2uB+1JqwYqVG3LIvdm9lj6imS/pQ==" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="/webjars/bootstrap/3.3.6/dist/css/bootstrap.min.css">
+    <script src="/webjars/bootstrap/3.3.6/dist/js/bootstrap.min.js"></script>
 
     <!-- react -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.14.2/react.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.14.2/react-dom.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.23/browser.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="css/my.css">
+    <script src="/webjars/react/15.1.0/react.min.js"></script>
+    <script src="/webjars/react/15.1.0/react-dom.min.js"></script>
+    <script src="/webjars/requirejs-babel/0.0.8/babel-5.8.22.min.js"></script>
+    <link rel="stylesheet" type="text/css" href="/css/my.css">
 </head>
 <body>
-    <!-- this navbar is just here for the looks, nothing cool happens in this index.html -->
-    <div class="navbar navbar-inverse navbar-static-top" role="navigation">
-        <div class="container-fluid">
-            <a class="navbar-brand" href="/">Destination Unknown</a>
-            <button class="navbar-toggle" data-toggle="collapse" data-target=".navHeaderCollapse">
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-            </button>
-            <div class="collapse navbar-collapse navHeaderCollapse">
-                <ul class="nav navbar-nav navbar-right">
-                    <li>
-                        <a href="/">Home</a>
-                    </li>
-                    <li>
-                        <a href="/about">About</a>
-                    </li>
-                    <li>
-                        <p class="navbar-text login-buttons-navbar">Login</p>
-                    </li>
-                </ul>
-            </div>
+<!-- this navbar is just here for the looks, nothing cool happens in this index.html -->
+<div class="navbar navbar-inverse navbar-static-top" role="navigation">
+    <div class="container-fluid">
+        <a class="navbar-brand" href="/">Destination Unknown</a>
+        <button class="navbar-toggle" data-toggle="collapse" data-target=".navHeaderCollapse">
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+            <span class="icon-bar"></span>
+        </button>
+        <div class="collapse navbar-collapse navHeaderCollapse">
+            <ul class="nav navbar-nav navbar-right">
+                <li>
+                    <a href="/">Home</a>
+                </li>
+            </ul>
         </div>
     </div>
-    <div id="container" >
-    </div>
-    <script type="text/babel" src="js/playfield.js"></script>
+</div>
+<div id="container">
+</div>
+<script type="text/babel" src="/js/playfield.js"></script>
 </body>
 </html>
